+++ Runtime/XRHandAnalyticsData.cs
@@ -1,4 +1,4 @@
-#if UNITY_EDITOR && ENABLE_CLOUD_SERVICES_ANALYTICS
+#if UNITY_EDITOR && (ENABLE_CLOUD_SERVICES_ANALYTICS || UNITY_2023_2_OR_NEWER)
 namespace UnityEngine.XR.Hands.Analytics
 {
     /// <summary>
@@ -17,4 +17,4 @@ namespace UnityEngine.XR.Hands.Analytics
         }
     }
 }
-#endif // UNITY_EDITOR && ENABLE_CLOUD_SERVICES_ANALYTICS
+#endif // UNITY_EDITOR && (ENABLE_CLOUD_SERVICES_ANALYTICS || UNITY_2023_2_OR_NEWER)