+++ CHANGELOG.md
@@ -8,6 +8,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [1.7.1] - 2025-10-24
+
+### Fixed
+- Fixed compiler issue with Editor analytics when using some platforms.
+- Fixed the XRHandShape Editor UI not updating the Hand Shape correctly in Play Mode using Link.
+- Fixed XRFingerShapeCondition throwing an error when using `XRFingerShapeType.Unspecified`.
+- Fixed XR Hand Capture issue where recording imports failed due to non-empty error stream.
+
 ## [1.7.0] - 2025-10-16
 
 ### Added